+++ packages/frontend/src/global.d.ts
@@ -73,5 +73,6 @@ declare global {
      */
     __internal_current_message_list_instance_id?: symbol
     __updateAccountListSidebar: (() => void) | undefined
+    __closeAllDialogs: () => void | undefined
   }
 }